split out GC tests to a new file, serialize them
This creates a new file, test-liveslots-real-gc.js, to house all the
tests from test-liveslots.js which depend on actual GC. We have known
but mysterious problems with consistency of garbage collection when
run under Node.js and AVA. Using test.serial() to inhibit parallel
execution of those tests seems to help.

In the long run, I want to have most of the liveslots GC code to be
tested with `makeMockGC` fixtures, which can give us more control over
precisely when collection and finalization happens, so we can exercise
more cases.

It also splits out a number of mock-GC tests, to
test-liveslots-mock-gc.js . These do not use engine-level GC, and are
properly deterministic.