improve error thrown during collection.get(missingRemotable)
Previously, doing a virtualCollection.get() on a Remotable that was
not present in the Map or Set would trigger an internal
assertion (complaining about a "missing ordinal"), rather than
reaching the usual error message (which includes the collection's
label as a debugging aid).

This fixes the `get()` path for both maps and sets to check for
ordinal-less Remotables first, before attempting to convert the key
into a dbKey. This let us signal the right error message.

It also updates some downstream tests which asserted the previous
behavior.

closes #7231