chore(swingset): simplify vat loader
For historical reasons, we had three distinct entrypoints into
vat-loader.js, for starting a new dynamic vat, restarting an existing
dynamic vat, and restarting an existing static vat. We now think of
vat creation/initialization as one thing, and launching a worker as a
different thing, so we can collapse all three into a single `create`
function.

vat-warehouse is now responsible for reacting to errors differently in
the three cases (errors in launching a worker for a
previously-established vat will panic the kernel, as do errors for
static vats).

The kernel passes 'panic' to makeVatWarehouse, instead of
makeVatLoader.