chore(swingset): RecordedVatOptions holds workerOptions, not managerType
The internal record of vat options is stored by the vatKeeper's
`setSourceAndOptions()`, and is of type RecordedVatOptions. This is
passed to the vat loader to decide how to construct ManagerOptions to
start the worker.

This commit replaces the `.managerType` property of RecordedVatOptions
with a new WorkerOptions record. For now, this only has a `.type`
field, whose value is either `local` or `xsnap`. Soon, the xsnap
flavor will acquire new `bundleIDs` properties.

The user-facing vat-options APIs (config.vats, kernel.createTestVat,
and E(vatAdminSvc).createVat) continue to use a `managerType` of
'local' or 'xs-worker' to control the selection.

(I prefer the name `xsnap` to describe the XS-in-subprocess worker,
but the old 'xs-worker' name is used extensively by unit tests in
other packages, and is impractical to change right now).

The `makeWorkerOptions()` helper function is used to convert the
user-facing options into the internal form.