liveslots: rewrite virtual object manager to use the new cache
This overhauls the VOM, to drastically simplify the creation of
virtual-object Representatives, and remove the syscall sensitivity to
GC.

The old code created `state` (with accessors for each state property)
and `context` (which holds both `state` and `self` or the `facets`
cohort) each time it needed to create a Representative, which is every
time a virtual/durable vref is deserialized but a Representative was
not present in the slotToVal table (thus sensitive to GC). This was
partially a holdover from the time when Representatives were behavior
methods closing over `state`.

Since then, we moved to a model where Representatives are empty
objects, with a special prototype that holds the behavior methods. We
use `defendPrototype` from `exo-tools.js` to build this prototype,
which wraps each behavior method with code that looks the
Representative up in a "context map" to find the `context`
object. This reduces memory usage: we share the prototype object among
all instances of the Kind, and the Representatives have no own
properties.

It also enables the key improvement of this commit: we can defer
creating `context` until userspace actually invokes a method on the
Representative. We replace the context map with a `contextProvider`: a
function that takes the Representative and looks up (or creates) a
context. Since userspace invocation of methods is unaffected by GC,
calls to the contextProvider are too, making it a safe place to
perform syscalls (such as the vatstoreGet needed to learn the state
properties and create the `state` accessors).

The cache ensures that we create the context at most once per
delivery.

We use two instances of the cache: one for `context` (including
`state`), and a second for the actual state data.

For each baseref, the state data cache holds the full `capdatas`
record (separate capdata for each property name). It also holds a
sparsely/lazily-populated `valueMap`, a Map of unserialized
values. These are made available to `state` getters. When modifying
a property on `state`, we immediately serialize it and hold it in the
cache. We also perform refcount delta management at that time. Later,
during flush(), we reassemble the capdatas (some modified, some
pristine) into the value of a `syscall.vatstoreSet`.

We defer unserialization until a getter is called, which reduces the
constraints on userspace behavior, and making virtual objects behave
more like virtual collections.

We remove `virtualObjectCacheSize` from vat options, since the new
cache size is unlimited, and thus not configurable.

This changes the `slotToVal` mapping to hold the cohort *record*,
where previously it held an Array. This is for multi-facet Kinds,
whose list of named facets is learned during `defineKindMulti`. Each
Kind has a fixed list of facet names, so each facet has both a name
and an index.

Userspace receives these as a record, with one property per
facet.

Internally, we need the index to encode into the vref, e.g. `o+11/5:1`
means the second (0-indexed) facet of the 5th (1-indexed) instance of
the Kind with KindID 11. We also use the index when looking up the
export status, which is recorded in one vatstore key per baseref, with
one character per facet, so 'rrn' means the first two facets are
reachable, and the third is unreachable.

Previously, during construction or reanimation, we also created an
*Array* of facets, and recorded them in the slotToVal Map (under the
baseref as a key). When doing `convertSlotToVal()`, we'd extract the
baseref from the vref, use it to fetch the array, then use the integer
facet index to extract the specific facet.

This complicated the construction process, because we needed to create
two separate containers for the set of facets, which could have
separate lifetimes. We needed a third object (named `keepAlive`) to
help maintain the cyclic GC-reference subgraph containing all facets
and the user-visible facet record.

With this commit, we discard this extra Array, and store the facet
record itself in slotToVal. `convertSlotToVal()` asks the VRM to
convert the facet index (extracted from the vref) into a facet name,
and then looks it up in the slotToVal record. When registering a
cohort, `registerValue` iterates through the facet names to make sure
we register each separate facet in the reverse-pointing valToSlot
table.

This gets rid of `toRegister`, `toExpose`, and `keepAlive`.