add helpful error when virtual-object initial data is a non-object
When building a fat-arrow function to construct a virtual-object's
initial data, it's easy to forget the parentheses, and wind up with a
visually-sensible but syntactically-incorrect initializer (and initial
data of `undefined`, so the resulting error is pretty obscure).

This adds a check in the VOM to catch this case and provide a better
error message.