test-terminate(-replay): reduce GC sensitivity
I saw a test failure here, "historical inaccuracy in replay", which
might have been due to GC happening one way in the original, and a
different way in the replay (when running under Node.js). This feels
like an aspect of #5575, and this test isn't trying to exercise
anything about GC, so I'm just going to set defaultReapInterval to
'never' to inhibit BOYD, which should remove the problem.