rename snapstore 'endPos' to 'snapPos'
This should reduce the chances of confusing it with
`transcriptSpan.endPos`. The two numbers are related: `snapPos` is the
same as `endPos` in the brief moment when we've just written out the
snapshot, but we immediately append a `save-snapshot` event, causing
`endPos` to increment. The `endPos` of the transcript span, being an
exclusive-right-edge kind of number, will be one larger than the last
item in the span, which means it will also be one larger than the
`snapPos`.