remove leftover makeMarshal from vatPowers
A long time ago, `makeMarshal` used an module-level WeakSet to keep
track of which objects had been registered with Far, which mean the
code doing the marshalling (i.e. liveslots) and the code providing the
marshalled values (i.e. userspace, calling Far) needed to get those
two pieces from the same place.

This was a problem for vats, because they bundle `@endo/marshal` or
`@endo/far` independently of liveslots' copy.

We no longer do that, but for some period, we thought it was useful
and important to give userspace vat code access to the same
`makeMarshal` that liveslots used. To facilitate that, we included a
reference in `vatPowers.makeMarshal`, which meant that the supervisors
had to provide a copy in `allVatPowers`.

We remove that from manager-local a while back, but forgot to remove it
from the xsnap worker.

This removes the remnants of makeMarshal from vatPowers in all manager
types.