fix(vats): loadVat/loadCriticalVat now accepts bundleIDs too
Previously, the loadVat/loadCriticalVat tools accepted just a `name`,
which was both used as the `bundleName` for the vat (referencing only
bundles that were baked into the kernel via `config.bundles`), and
also as the `{ name }` option (providing a description of the new
vat). It had no ability to use bundles that were referenced by
their (hash) `bundleID`, which means it could not use bundles
installed (with `controller.validateAndInstallBundle`) after launch.

This changes `makeLazyVatLoader()` to produce a loadVat which accepts
`(name, bundleRef)`, where the bundleRef is either `{ bundleName}`
(which uses `config.bundles[bundleName]`, via
`vatAdminSvc.createVatByName`), or `{ bundleID }` (which uses
post-launch installed bundles, via
`vatAdminSvc.getBundleCap(bundleID)` and `.createVat(bundleCap)`).

It is backwards compatible with callers who only supply the name,
which is then used for both the description and the config.bundles
lookup.

This enables post-launch proposals to use `vatLoader(name,
vatSourceRef)` to install vats using bundles that were installed
during the proposal process, rather than being limited to
pre-configured ones.